WIP refactor distance
