+++ src/skallel_stats/__init__.py
@@ -1,6 +1,3 @@
 # flake8: noqa
-from .api.distance import pairwise_distance
-
-
-from . import numpy_backend
-from . import dask_backend
+from .distance.api import pairwise_distance
+from . import distance
