--- src/skallel_stats/dask_backend/__init__.py
@@ -1,1 +1,0 @@
-from . import distance
